+++ app/code/community/Nosto/Tagging/Model/Meta/Product.php
@@ -21,7 +21,7 @@
  * @category  Nosto
  * @package   Nosto_Tagging
  * @author    Nosto Solutions Ltd <magento@nosto.com>
- * @copyright Copyright (c) 2013-2016 Nosto Solutions Ltd (http://www.nosto.com)
+ * @copyright Copyright (c) 2013-2017 Nosto Solutions Ltd (http://www.nosto.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -44,6 +44,7 @@ class Nosto_Tagging_Model_Meta_Product extends NostoProduct
      */
     public static $customizableAttributes = array(
         'gtin' => '_gtin',
+        'supplier_cost' => '_supplierCost',
     );
 
     /**