+++ app/code/community/Nosto/Tagging/Model/System/Config/Source/Attribute.php
@@ -21,19 +21,80 @@
  * @category  Nosto
  * @package   Nosto_Tagging
  * @author    Nosto Solutions Ltd <magento@nosto.com>
- * @copyright Copyright (c) 2013-2016 Nosto Solutions Ltd (http://www.nosto.com)
+ * @copyright Copyright (c) 2013-2017 Nosto Solutions Ltd (http://www.nosto.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
 /**
- * General purpose source model attributes
+ * Abstract source model class for product attributes
  *
  * @category Nosto
  * @package  Nosto_Tagging
  * @author   Nosto Solutions Ltd <magento@nosto.com>
  */
-class Nosto_Tagging_Model_System_Config_Source_Attribute
+abstract class Nosto_Tagging_Model_System_Config_Source_Attribute
 {
+
+    /**
+     * @var int the product attribute type id
+     */
+    const PRODUCT_TYPE_ATTRIBUTE_ID = 4;
+
+    /**
+     * @var string the form key for the value
+     */
+    const OPTION_KEY_VALUE = 'value';
+
+    /**
+     * @var string the form key for the label
+     */
+    const OPTION_KEY_LABEL = 'label';
+
+    /**
+     * Returns all available product attributes
+     *
+     * @param array $filters ['field_to_filter' => 'value']
+     * @return Mage_Catalog_Model_Resource_Product_Attribute_Collection
+     */
+    public function getProductAttributesCollection(array $filters=array())
+    {
+        $resourceModel = Mage::getResourceModel(
+            'catalog/product_attribute_collection'
+        );
+
+        if (is_array($filters) && !empty($filters)) {
+            foreach ($filters as $attribute => $value) {
+                $resourceModel->addFieldToFilter(
+                    $attribute,
+                    $value
+                );
+            }
+        }
+        $attributes = $resourceModel
+            ->addFieldToFilter(
+                'entity_type_id',
+                self::PRODUCT_TYPE_ATTRIBUTE_ID
+            )
+            ->setOrder(
+                'attribute_code',
+                Varien_Data_Collection::SORT_ORDER_ASC
+            );
+
+        return $attributes;
+    }
+
+    /**
+     * List of attributes that cannot be added to tags due to data type and
+     * Magento's internal processing of attributes
+     *
+     * @var array
+     */
+    public static $notValidCustomAttributes = array(
+        'group_price',
+        'tier_price',
+        'media_gallery',
+    );
+
     /**
      * Returns all available attributes
      *
@@ -41,8 +102,33 @@ class Nosto_Tagging_Model_System_Config_Source_Attribute
      */
     public function toOptionArray()
     {
-        /* @var Nosto_Tagging_Helper_Data $nostoHelper */
-        $nostoHelper = Mage::helper('nosto_tagging');
-        return $nostoHelper->getProductAttributeOptions();
+        $attributes = $this->getProductAttributes();
+        $attributeArray = array(
+            array(
+                self::OPTION_KEY_VALUE => 0,
+                self::OPTION_KEY_LABEL => 'None'
+            )
+        );
+        /** @var Mage_Eav_Model_Attribute $attribute */
+        foreach ($attributes as $attribute) {
+            $code = $attribute->getData('attribute_code');
+            if (in_array($code, self::$notValidCustomAttributes)) {
+                continue;
+            }
+            $label = $attribute->getData('frontend_label');
+            $attributeArray[] = array(
+                self::OPTION_KEY_VALUE => $code,
+                self::OPTION_KEY_LABEL => sprintf('%s (%s)', $code, $label)
+            );
+        }
+
+        return $attributeArray;
     }
+
+    /**
+     * Returns a collection of attributes defined by the child class
+     *
+     * @return Mage_Catalog_Model_Resource_Product_Attribute_Collection
+     */
+    abstract public function getProductAttributes();
 }